Compile and run natively in Noetic/Qt5
Updates to the build and configuration files to allow the STDR Simulator to work in Noetic and Qt5 have been created.  Minor alterations to code were needed to accommodate the different locations of some Qt header files.  Some code was altered to remove the accommodating UTF8 coding as it is now default.  Finally, an alteration was necessary to accommodate the new way timers in Qt work.

This update also fixes a bug in stdr_specifications.xml that prevented co2_sensors from working.