Dependencies corrected and co2_sensor fixed, all for Melodic
This is a branch that works for ROS Melodic.  The main updates are dependencies that have been added to the stdr_gui CMakeLists.txt and package.xml that allow rosdep to get everything necessary to compile the package.

It also includes a minor fix so that the co2_sensor works again.